+++ CHANGES.md
@@ -7,6 +7,8 @@ The current `master` branch is in a nightly state.
 Pending fixes:
 
 - Added `VGSX::setSaveDataDirectory` method
+- Added `VGSX::setConsoleCallback` method
+- `VGSX::setLogCallback` changes to `std::function`
 - Changed the master volume range: 0-256 -> 0-255
 
 ## Version 0.24.0 **(Disruptive)**